Add mmp_phys to uberblock
Define a structure for detecting multiple simultaneous imports of a
pool.  Add that structure to the end of the uberblock.

Initialize the structure when pools are imported or created, and
increment its' sequence count, mmp_seq, with each sync of a pool.

Extend zdb so that zdb -l <vdev> -u prints the mmp fields.