MMP: generate mmp_open_id only from spa_open_common()
Remove redundant call to mmp_init() from spa_create().  Initialization
is accomplished via spa_create()->spa_add()->mmp_init().

Set mmp_open_id in spa_open_common(), so it is set only upon import.

It is not necessary to set it in spa_create(); ZFS writes the empty pool
to stable storage and then immediately exits.  Before the initial write,
there is no pool on disk for another node to potentially import.  After
the initial write, the node which created the pool performs an import
and will follow the MMP protocol (when it's implemented).