MMP: reserve last 3 uberblocks in each label
For each ring (array of uberblocks in a label), reserve the last 3
uberblocks for use by MMP.

Make VDEV_UBERBLOCK_COUNT() return 3 less than the number of physical
slots in the ring.  This causes uberblock writes and reads not to
involve the last 3 uberblock slots used by MMP.

When checking for a foreign node writing to the pool, examine only the
reserved uberblocks in each label.

In zdb, dump the MMP information only for the reserved blocks.  Mark
them as such in the output.