Add delegations support
The following zpool/zfs commands should be runable as a normal
non-priviledged user.

  * zpool list|iostat|status|get
  * zfs list|get

Adding this functionality primarily required implementing the
secpolicy_* functions by mapping them to the equivilant Linux
capabilities.  Since the VFS already enforces several of these
checks it was not nessisary to implement them all.

In order to verify this functionality works as designed the ZFS
Test Suite cli_user/* and delegation/* tests have all been enabled.
These tests are run as a normal user and are used to verify
permissions are enforced.

The `zpool_create_001_neg` and `zpool_add_001_neg` test cases were
slightly modified for Linux.  They now allow a normal user to run
these commands in dry-run mode since no escalated priviledges are
required.

Two minor bug fixes were required for test-running.py.  First, the
Timer() object cannot be safely created in a `try:` block when there
is an unconditional `finally` block which references it.  Second,
when running as a normal user also check for scripts using the
both the .ksh and .sh suffixes.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #362