Implementation of vdev_raidz parity generate and reconstruct routines
- specialized gen/rec routines for all RAIDZ levels,
- new scalar raidz implementation (unrolled),
- two x86_64 SIMD implementations (SSE and AVX2 instructions sets),
- fastest routines selected on module load (benchmark).
- cmd/raidz_test - verify and benchmark all implementations against original

New zfs module parameters:
- zfs_vdev_raidz_impl (str): selects the implementation to use. On module load, the
parameter will only accept first 3 options, and other implementations can be set once
module is finished loading. Possible values for this option are:
    "fastest" - use the fastest math available
    "original" - use the original raidz code
    "cycle" - cycle trough all available impl for each new vdev_raidz
    "scalar" - new scalar impl
    "sse" - new SSE impl if available
    "avx2" - new AVX2 impl if available

see contents of `/sys/module/zfs/parameters/zfs_vdev_raidz_impl` to get list of supported
values. If an implementation is not supported on the system, it will not be showed. Currently
selected option is enclosed in `[]`.