[WIP] Distributed Parity (dRAID) Feature
WARNING: This is still work in progress.  The user interface
and on-disk format have changed from previous versions of this
PR.  It is not compatible with previous versions.  The on-disk
format is not finalized and may continue to change in future
versions.

This patch adds a new top-level vdev type called dRAID, which
stands for Distributed parity RAID.  This pool configuration
allows all dRAID vdevs to participate when rebuilding to a hot
spare device.  This can substantially reduce the total time
required to restore full parity to pool with a failed device.

Additionally, a new sequential rebuild mechanism has been
introduced which allows dRAID and mirror devices to be rebuilt
in LBA order.  This can be faster than a traditional rebuild
is the average block size is small.  However, it cannot verify
a blocks checksums as part of the rebuild so it's recommended
that a scrub be run after the rebuild completes.

A dRAID pool can be created using the new top-level `draid` type.
Like `raidz`, the desired redundancy is specified after the type:
`draid[1,2,3]`.  No additional information is required to create
the pool and reasonable default values will be chosen based on
the number of child vdevs in the dRAID vdev.

    zpool create <pool> draid[1,2,3] <vdevs...>

Unlike raidz, additional optional dRAID configuration values can
be provided as part of the draid type as colon separated values.
This allows administrators to fully specify a layout for either
performance or capacity reasons.  The supported options include:

  - draid[:<groups>g] - Redundancy groups
  - draid[:<spares>s] - Distributed hot spares (default 1)
  - draid[:<data>d]   - Data devices per group
  - draid[:<iter>i]   - Iterations perform when generating
                        a valid dRAID mapping (default 3)

As part of adding test coverage for the new dRAID vdev type
the following options were added to the ztest command.  These
options are leverages by the zloop.sh test script to test a
wide range of dRAID configurations.

  -K draid|raidz|random -- kind of RAID to test
  -D <value> -- dRAID data drives per redundancy group
  -G <value> -- dRAID redundancy group count
  -S <value> -- dRAID distributed spare drives
  -R <value> -- RAID parity (raidz or dRAID)
  -L         -- (Existing -G (dump log option) was renamed -L)

The process of updating the ZFS Test Suite to provide functional
coverage has been started.  The zpool_create and zpool_import
test groups have been extended, but all of the cli_root/zpool_*
groups still need to be considered.

TODO:
[x] - Rebased on master, will be frequently rebased from now on.
[x] - Add dRAID config validation functionality.
[x] - Enforced reasonable defaults to prevent harmful configs.
[x] - Move common dRAID functions to zcommon.
[x] - Replaced `draidcfg` command with `zpool create ...`.
[x] - Add functionalty to load/save known dRAID layouts.
[x] - Convert custom dRAID debugging to normal ZFS debugging.
[x] - Cleaned up 'zpool status' output.
[x] - Permutations for 255 device pool reduce to fit in label.
[x] - Rebuild works with virtual hot spare and physical device.
[x] - Allow adding new top-level dRAID vdevs (no removal).
[x] - Logical spares are now mandatory for dRAID.
[x] - Update `ztest` to add dRAID pools to its pool layouts.
[x] - User commands updated to detect dRAID kmod support.
[x] - Resolve checksum errors for non-uniform groups in pools
[ ] - Consider a verbose `zpool list --draid-groups` output.
[ ] - Investigate reducing the permutations size in the label.
[ ] - Add a utility to generate known balanced layouts.
[ ] - Generate known good dRAID layouts common configurations.
[ ] - Review and update the sequential rebuild code.
[ ] - Debug (or remove) dRAID mirror code (currently disabled).
[ ] - Add `zpool replace` option to request rebuild or resilver.
[ ] - Integrate with deferred resilver functionality.
[ ] - Add new and extend existing ZTS test cases.
[ ] - Investigate stale labels on disk preventing pool import.
[ ] - Verify checksum errors are reported correctly (zinject).
[ ] - Verify corruption repair works correctly.
[ ] - Support 'zpool detach/attach' for rebuild during sparing.
[ ] - Optimizate the dRAID allocation logic.
[ ] - Implement vdev_xlate() to support initialize and trim.
[ ] - Add support for ZED to kick in logical spares.
[x] - Extend zdb to omit or print permutation groups.
[ ] - Extend `zdb` for long options, not absolutely needed.
[ ] - Documentation updates (man pages, comments, wiki, etc).
[ ] - Performance optimizaiton / analysis.
[ ] - Warning, inflate block size, for zvols with small record size.
[ ] - Evaluate possible `draid.run` file for additional tests.
[ ] - Evaluate and explain space allocation.
[ ] - Evaluate the existing rebuild throttle
[ ] - Verify gang block handling works correctly.
[ ] - Verify rebuild/resilver `zpool status` reporting.
[ ] - Update packaging as needed.

Co-authored-by: Isaac Huang <he.huang@intel.com>
Co-authored-by: Mark Maybee <mmaybee@cray.com>
Co-authored-by: Don Brady <don.brady@delphix.com>
Co-authored-by: Srikanth N S <nsrikanth@cray.com>
Co-authored-by: Stuart Maybee <smaybee@cray.com>
Co-authored-by: Brian Behlendorf <behlendorf1@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>

External-issue: ZFS-12 ZFS-35 ZFS-36 ZFS-17 ZFS-56 ZFS 95 ZFS-96
External-issue: ZFS-100 ZFS-103 ZFS-106 ZFS-110 ZFS-111 ZFS-117
Issue #9558