Make ztest and zloop dRAID aware
Included new command options to control ztest testing with dRAID:
  -K draid|raidz|random -- kind of RAID to test
  -D <value> -- dRAID data drives per redundancy group
  -G <value> -- dRAID redundancy group count
  -S <value> -- dRAID distributed spare drives
  -R <value> -- RAID parity (raidz or dRAID)

The above values are used to generate a dRAID config from the
draidcfg command. This also allows for a more comprehensive set
of configurations when used with zloop.  By default, dRAID will
be provisioned 50% of the time in ztest runs.

Added dRAID specific section to zloop to exercise dRAID across a
wide range of configuration parameters. For example:

 ztest -VVVV -K draid -D 7 -G 6 -S 1 -m 0 -r 1 -R 1 -v 0 -a 12 -s 384m

Updated the zloop ztest time range to be 30 - 120 seconds.

Signed-off-by: Don Brady <don.brady@delphix.com>

Conflicts:
	cmd/ztest/ztest.c