Make remaining MMP activity check duration user-visible
MMP determines the duration of the import activity check based on
ub_mmp_delay and number of leaf vdevs.

If a pool has experienced I/O delays, so that ub_mmp_delay is large (ie
seconds), and the number of leaf vdevs is high, the activity check can
take tens of minutes or more.

This commit adds a kstat which reports the number of seconds remaining
in the activity check.  During a tryimport, this appears under

/proc/spl/kstat/zfs/$import/activity_check

and during an import, it appears under

/proc/spl/kstat/zfs/<poolname>/activity_check

When no activity check is occurring, it reports 0 seconds remaining.

This allows for automated systems such as an HA system to detect long
import times and abort, extend their timeouts, or log the condition.  It
will allow users to rule in/out activity check during an unexpectedly
long import.

A future PR can build on this to add a progress message during "zpool
import".

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>