MMP skip distributed spares
Originally from the ZFS-15 MMP commit abd5ab88.  dRAID distributed
spares should be skipped for MMP IOs.  This change should be
revisted since we may want to instead not add distributed spares
to the linked list of leaves.  But since this is known to work
it has been applied largely as is.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>