Fix allocation ASSERT
After accounting for alignment it's possible the allocation may
span metaslabs and trigger this ASSERT.  Convert the ASSERT in
to an error and let the normal allocation logic handle this.

Handling this case in the generic could would be preferable but
that may require further interface changes so for the moment
I've opted for the straight forward fix.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>