Use SCL_ZIO instead of SCL_STATE for mmp writes
MMP originally used vdev_label_write() for MMP writes, which required
it to take SCL_STATE as reader.

This is unnecessary for MMP, however.  After MMP issues a write, it
must be guaranteed that the vdev continues to exist; but the the vdev
state may safely change.  A failed or blocked MMP write due to a vdev
going offline does not prevent other MMP writes from being issued or
succeeding.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>